pyproject.toml
[build-system]
requires = ["setuptools>=64.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zilant-prime-core"
version = "0.1.0"
description = "Revolutionary Crypto-Core ZILANT Prime™: Secure container (.zil) with AEAD, Argon2id, and Proof-of-Sequential-Work barrier."
readme = "README.md"
license = { text = "MIT" }
authors = [
  { name = "ZILANT Prime Team" }
]
requires-python = ">=3.10,<4"
dependencies = [
  "cryptography>=42.0.0,<43.0.0",
  "argon2-cffi>=23.1.0,<24.0.0",
  "click>=8.1.0,<9.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0,<9.0.0",
  "pytest-cov>=5.0.0,<6.0.0",
  "flake8>=7.0.0,<8.0.0",
  "mypy>=1.8.0,<2.0.0",
  "ruff>=0.2.0,<0.3.0",
  "bandit>=1.7.0,<1.8.0",
  "hypothesis>=6.0.0,<7.0.0",
]

[project.scripts]
zilctl = "zilant_prime_core.cli:cli"

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = []

[tool.mypy]
strict = true

[tool.pytest.ini_options]
# сохраняем временные файлы в ./.pytest_temp и измеряем покрытие
addopts    = "--basetemp=.pytest_temp --cov=zilant_prime_core --cov-report=term-missing -q -s"
testpaths  = ["tests"]
